+++ apps/web/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Funding round setting to allow people who join during voting to get tokens and vote
 
+### Changed
+- Temporarily removed ability to customize email invite message, to prevent spam
+
 ## [6.5.6] - 2026-08-04
 
 ### Changed